+++ src/applications/appeals/10182/components/ShowIssuesList.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { getDate } from '../../shared/utils/dates';
 import { FORMAT_READABLE } from '../../shared/constants';
+import { getDate } from '../../shared/utils/dates';
 
 const ShowIssuesList = ({ issues }) => (
   <ul>
