+++ src/applications/appeals/10182/validations/issues.js
@@ -2,23 +2,7 @@ import { showExtensionReason } from '../utils/helpers';
 import { issueErrorMessages } from '../content/addIssue';
 import { content as extensionReasonContent } from '../content/extensionReason';
 
-import { maxSelectedErrorMessage } from '../../shared/content/contestableIssues';
 import { MAX_LENGTH, REGEXP } from '../../shared/constants';
-import { getSelected } from '../../shared/utils/issues';
-
-export const maxIssues = (
-  errors,
-  _fieldData,
-  formData,
-  _schema,
-  _uiSchema,
-  _index,
-  appStateData,
-) => {
-  if (getSelected(appStateData || formData).length > MAX_LENGTH.SELECTIONS) {
-    errors.addError(maxSelectedErrorMessage);
-  }
-};
 
 export const maxNameLength = (errors, data) => {
   if (data.length > MAX_LENGTH.NOD_ISSUE_NAME) {
