+++ src/applications/appeals/shared/content/errorMessages.js
@@ -1,4 +1,4 @@
-export const errorMessages = {
+export default {
   contestedIssue: 'You must select an eligible issue',
   missingIssue: 'You must add an issue',
   uniqueIssue: 'You must enter a condition you haven’t already entered',
